20/11/12

+++ WebContent/ServiceCenter/Q&Aboard/content.jsp
@@ -138,28 +138,33 @@
   <tr height="30">      
   
     <td colspan="4" align="right" > 
-    <%if("admin".equals(session.getAttribute("id"))){ %>
-   <input type="button" value="글수정" 
-       onclick="document.location.href='/ServiceCenter/Q&Aboard/qna_updateForm.jsp?num=<%=article.getNum()%>&pageNum=<%=pageNum%>'">
-      
-    &nbsp;&nbsp;&nbsp;&nbsp;
-   <input type="button" value="글삭제" 
-       onclick="document.location.href='/ServiceCenter/Q&Aboard/qna_deleteForm.jsp?num=<%=article.getNum()%>&pageNum=<%=pageNum%>'">
-     
-    &nbsp;&nbsp;&nbsp;&nbsp;
-    <%} %>
-       <input type="button" value="글목록" 
-       onclick="document.location.href='/ServiceCenter/Q&Aboard/Q&A.jsp?pageNum=<%=pageNum%>'">
-  <!--pagenum을  가지고 list.jsp로 넘어간다 -->
+   <%if("admin".equals(session.getAttribute("id"))){ %>
+    <%if(session.getAttribute("id").equals(article.getWriter())){%>  
+	   <input type="button" value="글수정" 
+	       onclick="document.location.href='/ServiceCenter/Q&Aboard/qna_updateForm.jsp?num=<%=article.getNum()%>&pageNum=<%=pageNum%>'">
+	 	  &nbsp;&nbsp;&nbsp;&nbsp;
+	   <input type="button" value="글삭제" 
+	       onclick="document.location.href='/ServiceCenter/Q&Aboard/qna_deleteForm.jsp?num=<%=article.getNum()%>&pageNum=<%=pageNum%>'">
+			&nbsp;&nbsp;&nbsp;&nbsp;
+			<input type="button" value="글목록" 
+	       onclick="document.location.href='/ServiceCenter/Q&Aboard/Q&A.jsp?pageNum=<%=pageNum%>'">
+	    	&nbsp;&nbsp;&nbsp;&nbsp; 
+		<%} %>
+		
+	      <input type="button" value="답글쓰기" 
+	       onclick="document.location.href='/ServiceCenter/Q&Aboard/qna_writeForm.jsp?num=<%=num%>&ref=<%=ref%>&re_step=<%=re_step%>&re_level=<%=re_level%>'">
+	      &nbsp;&nbsp;&nbsp;&nbsp;
+		<%} %>
+	       <input type="button" value="글목록" 
+	       onclick="document.location.href='/ServiceCenter/Q&Aboard/Q&A.jsp?pageNum=<%=pageNum%>'">
+	    	&nbsp;&nbsp;&nbsp;&nbsp; 
     </td>
   </tr>
-</table>    
+</table>  
+
 <%
  }catch(Exception e){} 
  %>
-</form>   
-</div>
-</div>
-</section>   
+</form>      
 </body>
 </html>
